Move utils inside lib
